roadsInHackerland challenge
The solution below uses Dijkstra algorithm which is an efficient approach
for finding the shortest path between nodes in a graph. Howver, it doesn't
fit very well for this problem. With increasing number of nodes, the algorithms
might take hours to find the answer if it doesn't crash.

Thus, this is committed just for fun. There should be a better algorithm out
there.